adds grouping to fixtures
